requirements.txt
Django==2.0
djangorestframework==3.10.2
entrypoints==0.3
mccabe==0.6.1
Pillow==6.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
